Renamed findNodeHandler stack/fiber RN wrapper
