+++ src/renderers/native/takeSnapshot.js
@@ -16,9 +16,9 @@ const UIManager = require('UIManager');
 
 import type {Element} from 'React';
 
-const findNodeHandle = ReactNativeFeatureFlags.useFiber
-  ? require('findNodeHandleFiberWrapper')
-  : require('findNodeHandleStackWrapper');
+const findNumericNodeHandle = ReactNativeFeatureFlags.useFiber
+  ? require('findNumericNodeHandleFiber')
+  : require('findNumericNodeHandleStack');
 
 /**
  * Capture an image of the screen, window or an individual view. The image
@@ -47,7 +47,7 @@ function takeSnapshot(
   },
 ): Promise<any> {
   if (typeof view !== 'number' && view !== 'window') {
-    view = findNodeHandle(view) || 'window';
+    view = findNumericNodeHandle(view) || 'window';
   }
 
   // Call the hidden '__takeSnapshot' method; the main one throws an error to
